ldap: Move `LDAPError` class to a utility module
