+++ salt/utils/ldap.py
@@ -1,0 +1,16 @@
+"""Common classes shared between LDAP execution and state modules."""
+
+
+class LDAPError(Exception):
+    """Base class of all LDAP exceptions raised by backends.
+
+    This is only used for errors encountered while interacting with
+    the LDAP server; usage errors (e.g., invalid backend name) will
+    have a different type.
+
+    :ivar cause: backend exception object, if applicable
+    """
+
+    def __init__(self, message, cause=None):
+        super().__init__(message)
+        self.cause = cause
